Put start/end in URL hash so they can be linked to
Clicking 'Go' now puts the start and end values in the location hash.

When the page is opened, the values are pulled out of the hash and the
inputs are set to the given values. If both start and end have values in
the hash, then it waits 1 second and starts trying to find a path.

`+` in the hash are correctly converted back into spaces.